Remove static function in Portalnetwork and create new constructor file
